Extract integration tests setup to own package
To make future PRs smaller.
