+++ integration/findandmodify_test.go
@@ -24,6 +24,7 @@ import (
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 
+	"github.com/FerretDB/FerretDB/integration/setup"
 	"github.com/FerretDB/FerretDB/integration/shareddata"
 )
 
@@ -139,7 +140,7 @@ func TestFindAndModifySimple(t *testing.T) {
 		name, tc := name, tc
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+			ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 			command := append(bson.D{{"findAndModify", collection.Name()}}, tc.command...)
 
@@ -170,7 +171,7 @@ func TestFindAndModifyEmptyCollectionName(t *testing.T) {
 		name, tc := name, tc
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+			ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 			var actual bson.D
 			err := collection.Database().RunCommand(ctx, bson.D{{"findAndModify", ""}}).Decode(&actual)
@@ -292,7 +293,7 @@ func TestFindAndModifyErrors(t *testing.T) {
 		name, tc := name, tc
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+			ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 			command := append(bson.D{{"findAndModify", collection.Name()}}, tc.command...)
 
@@ -405,7 +406,7 @@ func TestFindAndModifyUpdate(t *testing.T) {
 		name, tc := name, tc
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+			ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 			command := bson.D{{"findAndModify", collection.Name()}, {"query", tc.query}}
 			command = append(command, tc.command...)
@@ -544,7 +545,7 @@ func TestFindAndModifyUpsert(t *testing.T) {
 		name, tc := name, tc
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+			ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 			command := append(bson.D{{"findAndModify", collection.Name()}}, tc.command...)
 
@@ -588,7 +589,7 @@ func TestFindAndModifyUpsertComplex(t *testing.T) {
 		name, tc := name, tc
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+			ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 			command := append(bson.D{{"findAndModify", collection.Name()}}, tc.command...)
 
@@ -657,7 +658,7 @@ func TestFindAndModifyRemove(t *testing.T) {
 		name, tc := name, tc
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			ctx, collection := Setup(t, shareddata.Scalars)
+			ctx, collection := setup.Setup(t, shareddata.Scalars)
 
 			command := append(bson.D{{"findAndModify", collection.Name()}}, tc.command...)
