+++ integration/query_bitwise_test.go
@@ -25,12 +25,13 @@ import (
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 
+	"github.com/FerretDB/FerretDB/integration/setup"
 	"github.com/FerretDB/FerretDB/integration/shareddata"
 )
 
 func TestQueryBitwiseAllClear(t *testing.T) {
 	t.Parallel()
-	ctx, collection := Setup(t, shareddata.Scalars)
+	ctx, collection := setup.Setup(t, shareddata.Scalars)
 
 	// TODO: https://github.com/FerretDB/FerretDB/issues/508
 	_, err := collection.DeleteOne(ctx, bson.D{{"_id", "binary"}})
@@ -186,7 +187,7 @@ func TestQueryBitwiseAllClear(t *testing.T) {
 
 func TestQueryBitwiseAllSet(t *testing.T) {
 	t.Parallel()
-	ctx, collection := Setup(t, shareddata.Scalars)
+	ctx, collection := setup.Setup(t, shareddata.Scalars)
 
 	// TODO: https://github.com/FerretDB/FerretDB/issues/508
 	_, err := collection.DeleteOne(ctx, bson.D{{"_id", "binary"}})
@@ -310,7 +311,7 @@ func TestQueryBitwiseAllSet(t *testing.T) {
 
 func TestQueryBitwiseAnyClear(t *testing.T) {
 	t.Parallel()
-	ctx, collection := Setup(t, shareddata.Scalars)
+	ctx, collection := setup.Setup(t, shareddata.Scalars)
 
 	// TODO: https://github.com/FerretDB/FerretDB/issues/508
 	_, err := collection.DeleteOne(ctx, bson.D{{"_id", "binary"}})
@@ -458,7 +459,7 @@ func TestQueryBitwiseAnyClear(t *testing.T) {
 
 func TestQueryBitwiseAnySet(t *testing.T) {
 	t.Parallel()
-	ctx, collection := Setup(t, shareddata.Scalars)
+	ctx, collection := setup.Setup(t, shareddata.Scalars)
 
 	// TODO: https://github.com/FerretDB/FerretDB/issues/508
 	_, err := collection.DeleteOne(ctx, bson.D{{"_id", "binary"}})
