+++ integration/query_element_test.go
@@ -24,12 +24,13 @@ import (
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 
+	"github.com/FerretDB/FerretDB/integration/setup"
 	"github.com/FerretDB/FerretDB/integration/shareddata"
 )
 
 func TestQueryElementExists(t *testing.T) {
 	t.Parallel()
-	ctx, collection := Setup(t)
+	ctx, collection := setup.Setup(t)
 
 	_, err := collection.InsertMany(ctx, []any{
 		bson.D{{"_id", "empty-array"}, {"empty-array", []any{}}},
@@ -99,7 +100,7 @@ func TestQueryElementExists(t *testing.T) {
 func TestQueryElementType(t *testing.T) {
 	t.Parallel()
 	// TODO: add cases for "decimal" when it would be added.
-	ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+	ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 	for name, tc := range map[string]struct {
 		v           any
