+++ integration/query_test.go
@@ -25,12 +25,13 @@ import (
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 
+	"github.com/FerretDB/FerretDB/integration/setup"
 	"github.com/FerretDB/FerretDB/integration/shareddata"
 )
 
 func TestQueryUnknownFilterOperator(t *testing.T) {
 	t.Parallel()
-	ctx, collection := Setup(t, shareddata.Scalars)
+	ctx, collection := setup.Setup(t, shareddata.Scalars)
 
 	filter := bson.D{{"v", bson.D{{"$someUnknownOperator", 42}}}}
 	errExpected := mongo.CommandError{Code: 2, Name: "BadValue", Message: "unknown operator: $someUnknownOperator"}
@@ -42,7 +43,7 @@ func TestQuerySort(t *testing.T) {
 	t.Skip("TODO https://github.com/FerretDB/FerretDB/issues/457")
 
 	t.Parallel()
-	ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+	ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 	for name, tc := range map[string]struct {
 		sort        bson.D
@@ -170,7 +171,7 @@ func TestQuerySort(t *testing.T) {
 
 // TODO: https://github.com/FerretDB/FerretDB/issues/636
 func TestQuerySortValue(t *testing.T) {
-	ctx, collection := Setup(t, shareddata.Scalars)
+	ctx, collection := setup.Setup(t, shareddata.Scalars)
 
 	for name, tc := range map[string]struct {
 		sort        bson.D
@@ -310,7 +311,7 @@ func TestQuerySortValue(t *testing.T) {
 
 func TestQueryCount(t *testing.T) {
 	t.Parallel()
-	ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+	ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 	for name, tc := range map[string]struct {
 		command  any
@@ -363,7 +364,7 @@ func TestQueryCount(t *testing.T) {
 
 func TestQueryBadFindType(t *testing.T) {
 	t.Parallel()
-	ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+	ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 	for name, tc := range map[string]struct {
 		command bson.D
@@ -527,7 +528,7 @@ func TestQueryBadFindType(t *testing.T) {
 
 func TestQueryBadSortType(t *testing.T) {
 	t.Parallel()
-	ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+	ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 	for name, tc := range map[string]struct {
 		command    bson.D
@@ -589,7 +590,7 @@ func TestQueryBadSortType(t *testing.T) {
 func TestQueryExactMatches(t *testing.T) {
 	t.Parallel()
 	providers := []shareddata.Provider{shareddata.Scalars, shareddata.Composites}
-	ctx, collection := Setup(t, providers...)
+	ctx, collection := setup.Setup(t, providers...)
 
 	_, err := collection.InsertMany(ctx, []any{
 		bson.D{
@@ -647,7 +648,7 @@ func TestQueryExactMatches(t *testing.T) {
 
 func TestDotNotation(t *testing.T) {
 	t.Parallel()
-	ctx, collection := Setup(t)
+	ctx, collection := setup.Setup(t)
 
 	_, err := collection.InsertMany(
 		ctx,
@@ -719,7 +720,7 @@ func TestDotNotation(t *testing.T) {
 func TestQueryNonExistingCollection(t *testing.T) {
 	t.Parallel()
 
-	ctx, collection := Setup(t)
+	ctx, collection := setup.Setup(t)
 
 	cursor, err := collection.Database().Collection("doesnotexist").Find(ctx, bson.D{})
 	require.NoError(t, err)
