+++ integration/setup/setup.go
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package integration
+// Package setup provides integration tests setup helpers.
+package setup
 
 import (
 	"context"
