+++ integration/update_field_test.go
@@ -26,6 +26,7 @@ import (
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 
+	"github.com/FerretDB/FerretDB/integration/setup"
 	"github.com/FerretDB/FerretDB/integration/shareddata"
 	"github.com/FerretDB/FerretDB/internal/types"
 	"github.com/FerretDB/FerretDB/internal/util/testutil"
@@ -47,7 +48,7 @@ func TestUpdateFieldCurrentDate(t *testing.T) {
 		path := types.NewPathFromString("v")
 		result := bson.D{{"_id", id}, {"v", nowTimestamp}}
 
-		ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+		ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 		// store the current timestamp with $currentDate operator;
 		update := bson.D{{"$currentDate", bson.D{{"v", bson.D{{"$type", "timestamp"}}}}}}
@@ -247,7 +248,7 @@ func TestUpdateFieldCurrentDate(t *testing.T) {
 			name, tc := name, tc
 			t.Run(name, func(t *testing.T) {
 				t.Parallel()
-				ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+				ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 				res, err := collection.UpdateOne(ctx, bson.D{{"_id", tc.id}}, tc.update)
 				if tc.err != nil {
@@ -381,7 +382,7 @@ func TestUpdateFieldInc(t *testing.T) {
 			name, tc := name, tc
 			t.Run(name, func(t *testing.T) {
 				t.Parallel()
-				ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+				ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 				_, err := collection.UpdateOne(ctx, tc.filter, tc.update)
 				require.NoError(t, err)
@@ -471,7 +472,7 @@ func TestUpdateFieldInc(t *testing.T) {
 			name, tc := name, tc
 			t.Run(name, func(t *testing.T) {
 				t.Parallel()
-				ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+				ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 				_, err := collection.UpdateOne(ctx, tc.filter, tc.update)
 				require.NotNil(t, tc.err)
@@ -662,7 +663,7 @@ func TestUpdateFieldSet(t *testing.T) {
 		name, tc := name, tc
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+			ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 			res, err := collection.UpdateOne(ctx, bson.D{{"_id", tc.id}}, tc.update)
 			if tc.err != nil {
@@ -762,7 +763,7 @@ func TestUpdateFieldSetOnInsert(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
 
-			ctx, collection := Setup(t, shareddata.Composites)
+			ctx, collection := setup.Setup(t, shareddata.Composites)
 
 			opts := options.Update().SetUpsert(true)
 			var actualUpdateStat *mongo.UpdateResult
@@ -848,7 +849,7 @@ func TestUpdateFieldUnset(t *testing.T) {
 		name, tc := name, tc
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+			ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 			opts := options.Update().SetUpsert(true)
 			var actualStat *mongo.UpdateResult
@@ -913,7 +914,7 @@ func TestUpdateFieldMixed(t *testing.T) {
 		name, tc := name, tc
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			ctx, collection := Setup(t, shareddata.Scalars, shareddata.Composites)
+			ctx, collection := setup.Setup(t, shareddata.Scalars, shareddata.Composites)
 
 			opts := options.Update().SetUpsert(true)
 			var actualStat *mongo.UpdateResult
